feat: show issue pane for an issue upon LS request
Added a topic, which is published to from the LS middleware when a "snyk:" URI is recieved in a `window/showDocument` request with a "showInDetailPanel" "action" parameter. This topic is subscribed to by the `ProductServices`s which then display the relevant issue in the issue pane if the "product" parameter matches and the "issueId" is known.
The LS makes these requests for things like clicking a code action to request an AI auto fix.