feat: navigate in issue tree on code action click
Currently the code action needs to be pressed twice to trigger the LS to generate the AI code fix.
A new service, `ProductTreeViewService`, was created, which holds the `TreeView` and `ProductIssueTreeProvider` to allow for handling the navigation by getting the node from the `ProductIssueTreeProvider` and revealing it in the `TreeView`.
The `ProductIssueTreeProvider` was adapted to store the `TreeNode`s it generates, so they can be revealed by a call from this new service.

Fixed a bug where a `TreeNode`'s parent was never set due to a chicken and egg scenario.

Also removed remaining references to the feature being specific to AI fixes.
Also ran `npm run lint:fix`, which was previously missed.