feat: snyk uri handler [IDE-955] [IDE-975] (#580)
* feat: wip snyk uri handler

* chore: added comments

* feat: show issue pane for an issue upon LS request

Added a topic, which is published to from the LS middleware when a "snyk:" URI is recieved in a `window/showDocument` request with a "showInDetailPanel" "action" parameter. This topic is subscribed to by the `ProductServices`s which then display the relevant issue in the issue pane if the "product" parameter matches and the "issueId" is known.
The LS makes these requests for things like clicking a code action to request an AI auto fix.

* feat: navigate in issue tree on code action click

Currently the code action needs to be pressed twice to trigger the LS to generate the AI code fix.
A new service, `ProductTreeViewService`, was created, which holds the `TreeView` and `ProductIssueTreeProvider` to allow for handling the navigation by getting the node from the `ProductIssueTreeProvider` and revealing it in the `TreeView`.
The `ProductIssueTreeProvider` was adapted to store the `TreeNode`s it generates, so they can be revealed by a call from this new service.

Fixed a bug where a `TreeNode`'s parent was never set due to a chicken and egg scenario.

Also removed remaining references to the feature being specific to AI fixes.
Also ran `npm run lint:fix`, which was previously missed.

* fix: cleanup 1b7c5f6 leftovers

Remove leftover TODO.
Fix misuse of `sinon.fake()`.
Forgot to update CHANGELOG.md.

N.b. the code action does trigger the generate AI fix when 9533167 is applies on the LS.

* feat: make the AI fix apply button calls LS

Using the "snyk.code.fixApplyEdit" LS command along with the fix Id.
Renamed the message the details web pane sends to "fixApplyEdit", as "applyGitDiff" doesn't make sense anymore.
Removed all prior in IDE AI edit applying logic.

* fix: refactor per review comments

Stops using hard errors for bad showDocument URIs.

---------

Co-authored-by: Ben Durrans <Benjamin.Durrans@snyk.io>