docs: clarify pre-contribution asking
Updated CONTRIBUTING.md to clarify our expectations for changes and to ensure collaboration before work is done.
Updated the PR template to point to this and the CODE_OF_CONDUCT.md.
Changed the email address in the CODE_OF_CONDUCT.md to the new OSS Conduct Reports address.
Fixed a slight md issue in the Contributor-Agreement.md.