fix: inject clipboard and fix other tests [IDE-1538] (#682)
Inject the clipboard so we can do unit tests.
Plus used the window inject in a place that wasn't using it, but this required adding an overload for `showInformationMessage`.
Plus made the window interface use `Thenable` to more accurate the the VS Code interface.
Improve the type guard for `PresentableError`.
Fix the integration tests to test the new tree node layout.