test: use preview CLI when stable is not released
With a safeguard to ensure the toggle has been switched back to testing the stable CLI for the protocol version when the release pipeline runs.
Add a new test that ensures the stable CLI isn't available if the flag is set to say it isn't.
Made PR tests run on all PRs, not just those into main.
Run the PR CI pipelines of the branch, not rely on main's.